refactor: 특정 페이지 전용 컴포넌트 분리
