+++ src/components/index.ts
@@ -1,21 +1,7 @@
 import CartDescription from "./CartDescription/CartDescription";
 import CartList from "./CartList/CartList";
 import CartPrice from "./CartPrice/CartPrcie";
-import PaymentConfirm from "./PaymentConfirm/PaymentConfirm";
-import Header from "./common/Header/Header";
-import Footer from "./common/Footer/Footer";
-import DeliveryInfo from "./DeliveryInfo/DeliveryInfo";
-import CouponButton from "./CouponButton/CouponButton";
+import PaymentConfirm from "../pages/PaymentPage/components/PaymentConfirm/PaymentConfirm";
 import CouponList from "./CouponList/CouponList";
 
-export {
-  CartDescription,
-  CartList,
-  Header,
-  Footer,
-  CartPrice,
-  PaymentConfirm,
-  DeliveryInfo,
-  CouponButton,
-  CouponList,
-};
+export { CartDescription, CartList, CartPrice, PaymentConfirm, CouponList };
