+++ src/pages/OrderPage/components/DeliveryInfo/DeliveryInfo.tsx
@@ -1,8 +1,8 @@
 import { Wrapper, Info } from "./style";
-import { Text } from "../common";
-import { FilledCheckSvg, OutlineCheckSvg } from "../../assets";
+import { Text } from "../../../../components/common";
+import { FilledCheckSvg, OutlineCheckSvg } from "../../../../assets";
 import { useRecoilState } from "recoil";
-import { extremeDeliveryState } from "../../recoil";
+import { extremeDeliveryState } from "../../../../recoil";
 
 const DeliveryInfo = () => {
   const handleToggle = () => {
