+++ src/pages/OrderPage/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CouponButton } from "./CouponButton/CouponButton";
+export { default as DeliveryInfo } from "./DeliveryInfo/DeliveryInfo";
