--- src/pages/OrderPage/style.ts
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const SvgWrapper = styled.div`
-  cursor: pointer;
-`;
