+++ src/pages/PaymentPage/components/PaymentConfirm/style.ts
