+++ src/pages/PaymentPage/components/index.ts
@@ -1,0 +1,1 @@
+export { default as PaymentConfirm } from "./PaymentConfirm/PaymentConfirm";
